Added Doc_ExcludePat to register pathname suffixes you do not want found by DocFallback (like .bak and .swp files)
Eliminated fallback when the file name has nothing before
the suffix: like /mumble/.abc

The DocFallback code generates URL redirects to the new path
so that spelling errors like john.osterhoot are not
cached and propagated back to folks browsers.

Doc(errorUrl) is set after an error so that the /error.html
page template can use it.

Template caching traps the case where the cached .html file
cannot be written to disk.