New API that is designed to make integrating objects into Tclhttpd a bit easier.
This change adds
::Object_Url [virtual] [object] ?inthread 0?

Any urls under virtual are dispatched to the object to handle.

The API expects to see a method for the object called httpdMarshallArguments

httpdMarshallArguments [suffix] [cgiList]

Returns the actual command to execute to provide the url, or {} if the
combination of suffix and cgilist to not map to a valid object.